Crate zng_wgt_setup

Setup wizard widgets.

Crate

This crate is part of the zng project.

Modules

page

Common setup pages.

Structs

SetupWizard

W Represents a setup operation wizard.

Enums

SetupOp

Represents the operation a SetupWizard! is managing.

Statics

APP_ID_VAR

Unique identifier of the product being setup.

APP_NAME_VAR

Display name of the product being setup.

APP_ORG_VAR

Display name of the app producer.

APP_VERSION_VAR

Display version of the product being setup.

INSTALLED_VERSION_VAR

Display version of the product that is already installed.

SETUP_OP_VAR

Operation the setup wizard is managing.

Functions

app_id

P Defines the unique identifier of the product being setup.

app_name

P Defines the display name of the product being setup.

app_org

P Defines the display name of the producer of the product being setup.

app_version

P Defines the display version of the product being setup.

installed_version

P Defines the display version of the product that is already installed.

setup_op

P Defines what operation the wizard is configuring and running.